08-Speech2Text/requirements.txt
gradio
pydantic
python-docx
PyPDF2
openai
python-dotenv